requirements.txt
python-telegram-bot==13.15
requests==2.31.0
urllib3==1.26.18
certifi
chardet
idna
